style:Split the api into two files
